Add core game and editor target files for Aura project
This commit introduces the essential target files for the Aura project. These include both the game and editor target class files which are required for building the Aura game and the AuraEditor. It also sets up the base character class 'AuraCharacterBase', initializes the Aura module, and updates the '.uproject' file to include the new Aura module.