Add weapon attribute to AuraCharacterBase and implement EnhancedInput
A weapon attribute has been added to the AuraCharacterBase class, including its setup on the BeginPlay override. EnhancedInput has been implemented in Aura by adding it as a public dependency in Aura.Build.cs and creating the AuraPlayerController class. The new class object manages the input context and adds mapping on BeginPlay.